[#1630] Buttons replaced
Closes #1630
